Moved AddButton and related components to add-value folder
Also updated imports for certain components
